+++ vite.config.ts
@@ -1,4 +1,4 @@
-import { defineConfig } from 'vite';
+import { defineConfig } from 'vitest/config';
 
 export default defineConfig({
   build: {
@@ -6,5 +6,8 @@ export default defineConfig({
     emptyOutDir: true,
     target: 'es2022',
     cssCodeSplit: false,
+  },
+  test: {
+    exclude: ['**/node_modules/**', '**/target/**'],
   },
 });